Remove Parsl, switch to fully WDL with cromwell
The original motivation for wrapping with Parsl was to use the wqex
backend for automatic tuning of job resource requests according to their
actual usage. After playing around a bit, I think the RabbitMQ
Parsl-Cromwell connector is a bit too finicky for production. I'll have
to play around with it on a different project.